Bugfix: Fix matching of queries and responses when both IP addresses are the same
When source and destination IP addresses are the same, e.g. when testing
on localhost, query and response wouldn't always get matched. This was
because IP addresses and ports in DnsRecord structure were sorted
only by comparing IP addresses. This fix compares IP addresses and
ports separately so they are sorted correctly.