Increase burst delete limit in Transaction and TCP connections tables
Currently, burst delete limit for Transaction and TCP tables is
lower than the packet period after which timeout procedure is
triggered. This can lead the tables to gradually fill up
because they might not be able to timeout the records fast enough
if influx of bogus queries without responses is too large. That
in turn might limit probe's ability process traffic.

This commit sets burst delete limit for Transaction and TCP
connections tables to the same value as packet period for timeout
procedure (1024) which should give the tables the ability to timeout
records fast enough to not fill up.